refactor(pkg): split config_inspect into focused files
Break the monolithic config_inspect.go into smaller, focused files
for better maintainability and readability.

Closes #478
